requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
Flask-Cors==4.0.1
requests==2.31.0
mangum==0.19.0
asgiref==3.8.1
httpx==0.27.2
asyncio==3.4.3